test: add new feature to match utility
SPDK would like to use match in our project but need one specific
feature as our test tool output can contain a variable number of
irrelevant and platform specific information making the creation
of .match files very time consuming and with every new platform
we add to our test pool we'd need to update all match files.

The proposed changes here aren't needed for PMDK but a quick
review would be appreciated and obviously you are welcome to
merge them.

The idea is that the developer create an optional .ignore file
along with the .match file. The .ignore file contains one string
per line that the match program uses to effectively ignore any
comparisons on lines that have that string anywhere in them. The
updated match utility simply filters the match file and the output
file and removes any lines that have a substring of any of the
lines in the .ignore file.

Signed-off-by: paul luse <paul.e.luse@intel.com>